backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
python-multipart